+++ src/reloader.cpp
@@ -1,6 +1,7 @@
 #include "reloader.h"
 
 #include <algorithm>
+#include <cinttypes>
 #include <iostream>
 #include <ncurses.h>
 #include <thread>
@@ -184,7 +185,12 @@ void Reloader::reload_all(bool unattended)
 
 	t2 = time(nullptr);
 	dt = t2 - t1;
-	LOG(Level::INFO, "Reloader::reload_all: reload took %d seconds", dt);
+	// On GCC, `time_t` is `long int`, which is at least 32 bits. On x86_64,
+	// it's 64 bits. Thus, this cast is either a no-op, or an up-cast which are
+	// always safe.
+	LOG(Level::INFO,
+			"Reloader::reload_all: reload took %" PRId64 " seconds",
+			static_cast<int64_t>(dt));
 
 	const auto unread_feeds2 =
 		ctrl->get_feedcontainer()->unread_feed_count();
