+++ src/reloadthread.cpp
@@ -1,5 +1,6 @@
 #include "reloadthread.h"
 
+#include <cinttypes>
 #include <unistd.h>
 
 #include "logger.h"
@@ -14,8 +15,11 @@ ReloadThread::ReloadThread(Controller* c, ConfigContainer* cf)
 	, cfg(cf)
 {
 	LOG(Level::INFO,
-		"ReloadThread: waiting %u seconds between reloads",
-		waittime_sec);
+		"ReloadThread: waiting %" PRIi64 " seconds between reloads",
+		// In GCC, `time t` is `long int`, which is at least 32 bits. On
+		// x86_64, it's 64 bits. Thus, this cast is either a no-op, or an
+		// up-cast which are always safe.
+		static_cast<int64_t>(waittime_sec));
 }
 
 ReloadThread::~ReloadThread() {}
